+++ src/component/divider/index.hbs
@@ -1,0 +1,1 @@
+<div class='divider'></div>
