+++ src/component/field/index.hbs
@@ -1,0 +1,4 @@
+<div class="field">
+  <label for="{{name}}" class="field__label">{{label}}</label>
+  <input type="{{#if type}}{{type}}{{/if}}" value="{{#if value}}{{value}}{{/if}}" name="{{name}}" class="field__input">
+</div>
