+++ src/component/field/style.scss
@@ -1,0 +1,18 @@
+@use '../../lib/service.scss' as s;
+
+.field {
+    display: flex;
+    flex-direction: column;
+    gap: s.space(3);
+    width: 100%;
+
+    &__label {
+        font-size: s.size('text');
+    }
+
+    &__input {
+        padding: s.space(4) s.space(7);
+        border: 1px solid s.color('light');
+        border-radius: s.space(2);
+    }
+}
