+++ src/component/heading/index.hbs
@@ -1,0 +1,8 @@
+<section class='heading'>
+  {{#if title}}
+    <h1 class='title'>{{title}}</h1>
+  {{/if}}
+  {{#if description}}
+    <h2 class='description'>{{description}}</h2>
+  {{/if}}
+</section>
