+++ src/component/heading/style.scss
@@ -1,0 +1,30 @@
+@use '../../lib/service.scss' as s;
+
+.heading {
+    display: flex;
+    flex-direction: column;
+    gap: s.space(5);
+    margin-bottom: s.space(15);
+}
+.title {
+    font-size: s.size('title');
+    color: s.color('primary');
+    font-weight: bold;
+}
+.description {
+    font-size: s.size('textSubtitle');
+}
+
+.heading_products {
+    .heading {
+        align-items: center;
+    }
+    .title {
+        font-size: s.size('subTitle');
+        color: #000;
+        font-weight: bold;
+    }
+    .description {
+        font-size: s.size('textSubtitle');
+    }
+}
