+++ src/component/product-card/product-card.hbs
@@ -1,12 +1,10 @@
 <div class='product__card'>
   <div class='product__card__content'>
-    <h2 class='product__card__title'>{{name}}</h2>
+    <h2 class='product__card__title'>{{title}}</h2>
     <p
       class='product__card__description'
     >{{description}}</p>
-    <p class='product__card_id'>ID: {{id}}</p>
-  </div>
-  <div class='product__card__action'>
+
     <div class='product__card__price'>{{price}}$</div>
     <a
       href='/product-edit?id={{id}}'
