+++ src/component/purchase-product/index.hbs
@@ -1,0 +1,24 @@
+{{#with this}}
+  <div class='product__card'>
+    <img
+      src='{{img}}'
+      alt='{{title}}'
+      class='product-card__image'
+    />
+    <h2 class='product__card__title'>{{title}}</h2>
+    <p
+      class='product__card__description'
+    >{{description}}</p>
+
+    {{#each category}}
+      <span
+        class='product__card__category product__card__category--{{id}}'
+      >{{text}}</span>
+    {{/each}}
+    <a
+      href='/purchase-product?id={{id}}'
+      class='product__card__button'
+    >Купити зараз:
+      {{price}}$</a>
+  </div>
+{{/with}}
