+++ src/container/purchase-index/index.hbs
@@ -1,0 +1,22 @@
+{{! 
+Це код container файлу для index сторінки, 
+в якому міститься код, котрий буде віддаватися по шляху (PATH) - "/"
+}}
+
+<main class='main'>
+  {{> heading}}
+  <section class='purchase-list'>
+    <div class='product-cards'>
+      {{#each data.products}}
+      
+        {{> purchase-product data=this}}
+        
+      {{/each}}  
+    </div>
+  </section>
+</main>
+
+{{! 
+    Цей код містить лише те, що буде виводитись в <body>
+    Адже наш layout/default.hbs має вже структуру HTML
+}}
