+++ src/container/purchase-product/style.scss
@@ -1,0 +1,64 @@
+@use '../../lib/service.scss' as s;
+
+@use '../../component/heading/style.scss' as heading;
+@use '../../component/purchase-product/style.scss' as purchase;
+@use '../../component/divider/style.scss' as divider;
+@use '../../component/field/style.scss' as field;
+
+.main {
+    display: flex;
+    flex-direction: column;
+    gap: s.space(10);
+    padding: s.space(10);
+    max-width: 1200px;
+}
+
+.product-cards {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: s.space(15) s.space(5);
+}
+
+.product__form {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: s.space(5);
+    margin-bottom: s.space(15);
+}
+
+.product__info {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    gap: s.space(6);
+
+    &__title {
+        font-size: s.size('titleProduct');
+        color: s.color('primary');
+        font-weight: bold;
+        line-height: 120%;
+    }
+
+    &__description {
+        font-size: s.size('textSubtitle');
+        line-height: 150%;
+    }
+
+    &__image {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+
+
+    &__button {
+        width: fit-content;
+        background-color: s.color('primary');
+        color: s.color('secondary');
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        padding: s.space(4) s.space(10);
+        border-radius: s.space(2);
+    }
+}
