+++ src/container/user-success-info/index.hbs
