+++ src/layout/default/style.scss
@@ -3,4 +3,5 @@
 body {
     font-family: 'Roboto', sans-serif;
     max-width: 1000px;
+    margin: 0 auto;
 }
